Better ResizeLogic
Adds a Resize Observer to our element.

Ok bit of a long talk here.
When we call many elements of p5 or q5 and they are in div boxes that may shrink or grow, if we set the width of each on setup the sizes can change based on css rules and other page timings. With the way p5 works there is certain cases where you have to call resize 2 times to make sure that it fills the full div box. Using ResizeObserver we can make sure that all the resize calls not just ones post load are called. This does mean that we may see 2 calls to window resize on a resize but the only way to avoid  would be logic to check if the resize size and last resize size are the same, which may cause problems. In any case this helps solve all resize issues due to page size on setup in a flex box.

Thanks