More listener after init
If you load a sketch while holding a key or mouse. Or while moving quickly, you can get an error of Uncaught TypeError: this._mouseMovedFn is not a function.

This is because the listener can call before the function is declared. By moving the listener to the end of the init chain we make sure that all methods are initialized before calling.