Fix prior mouse updating.
This would incorrectly update the mouse to move based on the frame updates and not our draw events.
You can test this using a simple draw loop.
make a var that stores the last mouse pos, and compare it to the pmouse.
They do not equal each other.
The behavior in normal p5 is that they equal each other.